finished supporting procs, WIP history

+++ src/myshell.c
@@ -12,27 +12,74 @@
 #define TERMINATED -1
 #define RUNNING 1
 #define SUSPENDED 0
-    
+#define HISTLEN 20
+#define INPUT_MAX 2048
 
-typedef struct process{
-    cmdLine* cmd;                     /* the parsed command line*/
-    pid_t pid; 		                  /* the process id that is running the command*/
-    int status;                       /* status of the process: RUNNING/SUSPENDED/TERMINATED */
-    struct process *next;	          /* next process in chain */
+typedef struct process
+{
+    cmdLine *cmd;         /* the parsed command line*/
+    pid_t pid;            /* the process id that is running the command*/
+    int status;           /* status of the process: RUNNING/SUSPENDED/TERMINATED */
+    struct process *next; /* next process in chain */
 } process;
 
-void addProcess(process** process_list, cmdLine* recievedCmd, pid_t recievedPid)
+cmdLine * cmdSemiCopy(cmdLine *origin)
 {
-    struct process *toAdd = (struct process *) malloc(sizeof(struct process));
-    toAdd->cmd = recievedCmd;
-    toAdd->next = NULL;
-    toAdd->pid = recievedPid;
-    toAdd->status = RUNNING;
-    
-    (*process_list)->next = toAdd;
+    cmdLine *copy = (cmdLine *)malloc(sizeof(cmdLine));
+    char *prompt = malloc(32);
+    ((char**)copy->arguments)[0] = strcpy(prompt, origin->arguments[0]);
+    return copy;
 }
 
-void removeProcess(process **process_list, pid_t toRemove)
+process * makeProcess(cmdLine *recievedCmd, pid_t recievedPid)
+{
+    process *newProc = (struct process *)malloc(sizeof(struct process));
+    newProc->cmd = cmdSemiCopy(recievedCmd);
+    newProc->next = NULL;
+    newProc->pid = recievedPid;
+    newProc->status = RUNNING;
+    return newProc;
+}
+
+void addProcess(process **process_list, cmdLine *recievedCmd, pid_t recievedPid)
+{
+    process *toAdd = makeProcess(recievedCmd, recievedPid);
+    if (!(*process_list))
+    {
+        *process_list = toAdd;
+        return;
+    }
+
+    process *current = *process_list;
+    while (current->next)
+    {
+        current = current->next;
+    }
+        current->next = toAdd;
+}
+
+char *intToStatus(int status)
+{
+    char *statusString = malloc(20 * sizeof(char));
+    switch (status)
+    {
+    case -1:
+        strcpy(statusString, "Terminated");
+        break;
+    case 0:
+        strcpy(statusString, "Suspended");
+        break;
+    case 1:
+        strcpy(statusString, "Running");
+        break;
+    default:
+        strcpy(statusString, "Unknown");
+        break;
+    }
+    return statusString;
+}
+
+void removeProcessFromList(process **process_list, pid_t toRemove)
 {
     if (!(*process_list))
     {
@@ -42,22 +89,20 @@ void removeProcess(process **process_list, pid_t toRemove)
     process *prev = *process_list, *current = (*process_list)->next;
     if (prev->pid == toRemove)
     {
-        free(prev);
         *process_list = current;
         return;
     }
 
     while (current)
     {
+        process *next_process = current->next;
         if (current->pid == toRemove)
         {
-            process *next_process = current->next;
-            free(current);
             prev->next = next_process;
-            break;
+            break; /*don't worry, it is freed from outside this function!*/
         }
         prev = current;
-        current = current->next;
+        current = next_process;
     }
 }
 
@@ -71,9 +116,9 @@ void freeProcessList(process **process_list)
     }
 }
 
-void changeProcessStatus(process *process_list, pid_t toChange, int new_status)
+void updateProcessStatus(process **process_list, pid_t toChange, int new_status)
 {
-    process *current = process_list;
+    process *current = *process_list;
     while (current)
     {
         if (current->pid == toChange)
@@ -84,9 +129,81 @@ void changeProcessStatus(process *process_list, pid_t toChange, int new_status)
     }
 }
 
-void printProcessList(process** process_list)
+void updateProcessList(process **process_list)
 {
+    process *current = *process_list;
+    while (current)
+    {
+        int currentStatus;
+        int result = waitpid(current->pid, &currentStatus, WNOHANG | WUNTRACED);
+        if (result == 0)
+        {
+            currentStatus = RUNNING;
+        }
+        else
+        {
+            if (WIFSTOPPED(currentStatus))
+            {
+                currentStatus = SUSPENDED;
+            }
+            else if (WIFCONTINUED(currentStatus))
+            {
+                currentStatus = RUNNING;
+            }
+            else if (WIFEXITED(currentStatus) || WIFSIGNALED(currentStatus))
+            {
+                currentStatus = TERMINATED;
+            }
+        }
 
+        if (currentStatus != current->status)
+        {
+            updateProcessStatus(process_list, current->pid, currentStatus);
+        }
+        current = current->next;
+    }
+}
+
+void printProcess(process *proc)
+{
+    char *status = intToStatus(proc->status);
+    printf("%-*d %-*s   %-*s\n", 8, proc->pid,
+           8, proc->cmd->arguments[0],
+           8, status);
+    free(status);
+}
+
+process *removeProcess(process **process_list, process * procToRemove)
+{
+    removeProcessFromList(process_list, procToRemove->pid);
+    process *next_proc = procToRemove->next;
+    freeCmdLines(procToRemove->cmd); /*it's a copy anyway*/
+    free(procToRemove);
+    return next_proc;
+}
+
+void printProcessListAndDeleteIfTerminated(process **process_list)
+{
+    process *current = *process_list;
+    while (current)
+    {
+        printProcess(current);
+        if (current->status == TERMINATED)
+        {
+            current = removeProcess(process_list, current);
+        }
+        else
+        {
+            current = current->next;
+        }
+    }
+}
+
+void onProcs(process **process_list)
+{
+    printf("%-*s %-*s   %-*s\n", 8, "PID", 8, "Command", 8, "STATUS");
+    updateProcessList(process_list);
+    printProcessListAndDeleteIfTerminated(process_list);
 }
 
 char containsDebugFlag(int argc, char const *argv[])
@@ -99,7 +216,7 @@ char containsDebugFlag(int argc, char const *argv[])
     return 0;
 }
 
-int handleSpecialCommands(cmdLine *cmd, process *process_list)
+int handleSpecialCommands(cmdLine *cmd, process **process_list)
 {
     if (strcmp(cmd->arguments[0], "cd") == 0)
     {
@@ -109,29 +226,72 @@ int handleSpecialCommands(cmdLine *cmd, process *process_list)
     }
     if (strcmp(cmd->arguments[0], "suspend") == 0)
     {
-        int toStop = atoi(cmd->arguments[1]);
-        kill(toStop, SIGTSTP);
-        changeProcessStatus(process_list, toStop, SUSPENDED);
-        usleep(10000);
+        if (cmd->argCount > 1)
+        {
+            int toStop = atoi(cmd->arguments[1]);
+            kill(toStop, SIGTSTP);
+            usleep(10000);
+        }
         return 1;
     }
     if (strcmp(cmd->arguments[0], "wake") == 0)
     {
-        int toContinue = atoi(cmd->arguments[1]);
-        kill(toContinue, SIGCONT);
-        changeProcessStatus(process_list, toContinue, RUNNING);
-        usleep(10000);
+        if (cmd->argCount > 1)
+        {
+            int toContinue = atoi(cmd->arguments[1]);
+            kill(toContinue, SIGCONT);
+            usleep(10000);
+        }
         return 1;
     }
     if (strcmp(cmd->arguments[0], "kill") == 0)
     {
-        int toTerminate = atoi(cmd->arguments[1]);
-        kill(toTerminate, SIGINT);
-        changeProcessStatus(process_list, toTerminate, TERMINATED);
+        if (cmd->argCount > 1)
+        {
+            int toTerminate = atoi(cmd->arguments[1]);
+            kill(toTerminate, SIGINT);
+            usleep(10000);
+        }
+        return 1;
+    }
+    if (strcmp(cmd->arguments[0], "procs") == 0)
+    {
+        onProcs(process_list);
         usleep(10000);
         return 1;
     }
+
     return 0;
+}
+
+void freeHistory(char** history) {
+    for (int i = 0; i < HISTLEN; i++) {
+        free(history[i]);
+    }
+}
+
+void printHistory(char **history, int newest, int oldest)
+{
+    for (int i = newest; i < oldest; i++) {
+        printf(history[i]);
+    }
+}
+
+void historyCommands(char **history, int newest, int oldest, cmdLine *cmd)
+{
+    char *command = cmd->arguments[0];
+    if (command == "history")
+    {
+        printHistory(history, newest, oldest);
+    }
+    if (command == "!!")
+    {
+        
+    }
+    if (command == "!n")//this is wrong (n is a number)
+    {
+        
+    }
 }
 
 void execute(cmdLine *pCmdLine)
@@ -143,22 +303,34 @@ void execute(cmdLine *pCmdLine)
     }
 }
 
+
+//currently on: making sure printHistory works, implementing !!, !n
 int main(int argc, char const *argv[])
 {
     struct process *processList;
     char debug = containsDebugFlag(argc, argv);
+    char *history[HISTLEN];
+    int newest = -1, oldest = -1;
 
     while (1)
     {
-        char buffer[PATH_MAX], input[2048];
+        char buffer[PATH_MAX], input[INPUT_MAX];
         getcwd(buffer, PATH_MAX);
         printf("~%s$ ", buffer);
         fgets(input, 2048, stdin);
 
         if (strcmp(input, "quit\n") == 0)
+        {
+            freeProcessList(&processList);
+            freeHistory(history);
             break;
+        }
         if (strcmp(input, "\n") == 0)
             continue;
+
+        char *history_line = malloc(INPUT_MAX);
+        strcpy(history_line, input);
+        history[++oldest] = history_line;
 
         cmdLine *cmd;
         cmd = parseCmdLines(input);
@@ -166,7 +338,7 @@ int main(int argc, char const *argv[])
         if (debug == 1)
             printf("Executing: %s", input);
 
-        if (handleSpecialCommands(cmd, processList))
+        if (handleSpecialCommands(cmd, &processList))
         {
             freeCmdLines(cmd);
             continue;
@@ -181,7 +353,7 @@ int main(int argc, char const *argv[])
             {
                 freeCmdLines(cmd);
                 perror("Piping unsuccessful");
-                return 1;
+                break;
             }
         }
 
@@ -204,27 +376,32 @@ int main(int argc, char const *argv[])
             {
                 freopen(cmd->outputRedirect, "w", stdout);
             }
-            addProcess(&processList, cmd, pid1);
-            execute(cmd);
+            //correct the !n --------------------------------------------------------
+            if (cmd->arguments[0] == "history" || cmd->arguments[0] == "!!" || cmd->arguments[0] == "!n")
+            {
+                historyCommands(history, newest, oldest, cmd);
+            }
+            else
+                execute(cmd);
         }
         else
         { /*Main process*/
-            if (cmd->blocking == 1)
+            addProcess(&processList, cmd, pid1);
+
+            if (cmd->blocking)
             {
                 waitpid(pid1, NULL, 0);
-                removeProcess(&processList, pid1);
             }
             else
             {
                 usleep(10000);
             }
-                
 
             if (debug == 1)
             {
                 printf("Child PID1: %d\n", pid1);
             }
-           
+
             if (needToPipe)
             {
                 pid_t pid2 = fork();
@@ -234,12 +411,13 @@ int main(int argc, char const *argv[])
                     {
                         printf("Child PID2: %d\n", pid2);
                     }
+                    addProcess(&processList, cmd->next, pid2);
                     close(fd[1]);
                     close(fd[0]);
-                    waitpid(pid1, NULL, 0);
-                    waitpid(pid2, NULL, 0);
-                    freeCmdLines(cmd);
-                    usleep(10000);
+                    if (cmd->next->blocking)
+                    {
+                        waitpid(pid2, NULL, 0);
+                    }
                 }
                 else
                 { /*Child 2 process*/
@@ -254,11 +432,17 @@ int main(int argc, char const *argv[])
                     {
                         freopen(cmd->next->outputRedirect, "w", stdout);
                     }
-                    addProcess(&processList, cmd->next, pid2);
-                    execute(cmd->next);
+                    //correct the !n --------------------------------------------------------
+                    if (cmd->next->arguments[0] == "history" || cmd->next->arguments[0] == "!!" || cmd->next->arguments[0] == "!n")
+                    {
+                        historyCommands(history, newest, oldest, cmd->next);
+                    }
+                    else
+                        execute(cmd->next);
                 }
             }
-
+            usleep(10000);
+            freeCmdLines(cmd);
         }
     }
     return 0;
